refactor(cleanup): consolidate OrgNode into @paperclipai/shared (pass 6)
Moved OrgNode from server/src/routes/org-chart-svg.ts and ui/src/api/agents.ts
into @paperclipai/shared/src/types/agent.ts with collapsedReports marked
optional (server-only field, undefined in the UI view).

Both definition sites now import + re-export from shared so downstream
consumers (server/routes/agents.ts, server/services/company-portability.ts,
ui/pages/Agents.tsx, ui/pages/OrgChart.tsx, scripts/generate-org-chart-
satori-comparison.ts) need no change.

Skipped scripts/generate-org-chart-images.ts — it defines an unrelated
OrgNode interface with a different shape (children[] instead of reports[]).

Typecheck: all 20 workspace packages green.
