+++ packages/shared/src/index.ts
@@ -283,6 +283,7 @@ export type {
   AdapterEnvironmentTestStatus,
   AdapterEnvironmentCheck,
   AdapterEnvironmentTestResult,
+  OrgNode,
   AssetImage,
   Project,
   ProjectCodebase,
