+++ packages/shared/src/types/agent.ts
@@ -116,6 +116,16 @@ export interface AdapterModel {
   label: string;
 }
 
+export interface OrgNode {
+  id: string;
+  name: string;
+  role: string;
+  status: string;
+  reports: OrgNode[];
+  /** Populated by collapseTree on the server: flattened list of hidden descendants for avatar grid rendering. Server-only, undefined on UI. */
+  collapsedReports?: OrgNode[];
+}
+
 export type AdapterEnvironmentCheckLevel = "info" | "warn" | "error";
 export type AdapterEnvironmentTestStatus = "pass" | "warn" | "fail";
